+++ src/constants.ts
@@ -1,1 +1,8 @@
+import * as vscode from 'vscode';
+
 export const EXTENSION_PREFIX = 'formatto';
+
+export const DOCUMENT_SELECTOR: readonly vscode.DocumentFilter[] = [
+  { language: 'ruby' },
+  { language: 'gemfile' },
+];